engine(db): split seed_default_config into config_seeds/*.cpp per category
seed_default_config was 990 of database.cpp's 4744 lines: 63 upsert_config
calls in one function, grouped only by the order they were written in.
Move each category's entries verbatim into its own translation unit under
engine/src/db/config_seeds/, over a shared ConfigSeeder (config_seeds/seed.hpp)
that carries the upsert/preserve-value logic, the metadata-marker helpers and
the retired-key list. Database::seed_default_config is now the ordered list
of the seven calls.

ConfigSeeder also closes a gap the single function made impossible by
inspection: it refuses a key seeded twice across two category files, since
nothing else would catch that once the seed is split.

No behaviour change - every ConfigEntry literal and its wrapping metadata
markers moved unedited (verified with a scripted whitespace-insensitive diff
against the pre-split function).
