+++ engine/src/db/database.cpp
@@ -49,14 +49,13 @@
 #include <mutex>
 #include <sstream>
 #include <string_view>
-#include <thread>
 #include <unordered_map>
 #include <unordered_set>
 
+#include "config_seeds/seed.hpp"
 #include "vayu/core/constants.hpp"
 #include "vayu/core/spec_binding.hpp"
 #include "vayu/http/default_headers.hpp"
-#include "vayu/http/transport_policy.hpp"
 #include "vayu/utils/id.hpp"
 #include "vayu/utils/invariant.hpp"
 #include "vayu/utils/logger.hpp"
@@ -3743,1053 +3742,42 @@ double Database::get_config_double (const std::string& key, double default_value
     }
 }
 
-namespace {
-// Serialize all_http_versions() to the `{value,label}` JSON array the
-// "defaultHttpVersion" config entry stores in its `options` column. Derived
-// from the domain enumeration rather than a literal list, so a change to
-// HttpVersion cannot silently drift out of sync with the seeded options.
-std::string http_version_options_json () {
-    nlohmann::json options = nlohmann::json::array ();
-    for (const auto version : vayu::all_http_versions ()) {
-        options.push_back ({ { "value", vayu::to_string (version) },
-        { "label", vayu::http_version_label (version) } });
-    }
-    return options.dump ();
-}
-
-// The `{value,label}` array for "dbSynchronous". SQLite's synchronous levels
-// are its own enumeration, not ours, so the list is literal - but it belongs in
-// the `options` column rather than spelled out as "0 = Off, 1 = ..." prose over
-// an integer input, which is what the entry used to do.
-// The `{value,label}` array for "proxyMode". Derived from the enumeration the
-// resolver parses, so a mode added to `ProxyMode` cannot be missing from the
-// options the config route validates against - which would make it a value
-// the engine understands and `POST /config` refuses.
-std::string proxy_mode_options_json () {
-    nlohmann::json options = nlohmann::json::array ();
-    for (const auto mode : vayu::http::all_proxy_modes ()) {
-        options.push_back ({ { "value", vayu::http::to_string (mode) },
-        { "label", vayu::http::proxy_mode_label (mode) } });
-    }
-    return options.dump ();
-}
-
-std::string log_level_options_json () {
-    const nlohmann::json options = { { { "value", "debug" }, { "label", "Debug" } },
-        { { "value", "info" }, { "label", "Info" } },
-        { { "value", "warn" }, { "label", "Warning" } },
-        { { "value", "error" }, { "label", "Error" } } };
-    return options.dump ();
-}
-
-std::string db_synchronous_options_json () {
-    const nlohmann::json options = { { { "value", "0" }, { "label", "Off" } },
-        { { "value", "1" }, { "label", "Normal" } },
-        { { "value", "2" }, { "label", "Full" } } };
-    return options.dump ();
-}
-} // namespace
-
 void Database::seed_default_config () {
     std::lock_guard<std::recursive_mutex> lock (impl_->mutex);
 
-    // One transaction for the whole seed, for two reasons that happen to agree
-    // (issue #838).
-    //
-    // Correctness: this writes upwards of sixty rows - thirteen deletions of
-    // retired keys and one upsert per live entry - and each was its own
-    // implicit transaction, so a throw partway through left the config table
-    // half-seeded. The next start repaired it (a missing key is re-seeded, an
-    // existing one keeps the user's value), but "repaired on the next start"
-    // is not the same promise as "never observed half-written", and the
-    // config route can read this table while startup is still writing it.
-    //
-    // Cost: sixty-odd commits became one. On a scratch database that is the
-    // difference between sixty WAL commits per test process and a single one,
-    // which is what makes this show up in a test-suite wall time at all. The
-    // guard rolls back if anything below throws; `commit` is at the end.
+    // One transaction for the whole seed (issue #838): a throw partway used to
+    // leave the table half-seeded, and batching what were ~66 separate commits
+    // into one is what keeps a scratch database's test-suite wall time down.
     auto seed_transaction = impl_->storage.transaction_guard ();
 
-    // Retired settings: keys nothing reads any more. Delete any row left behind
-    // by an older version so the Settings UI (which renders engine entries
-    // dynamically from GET /config) stops offering a dead knob.
-    //
-    //   requestBatchSize - drove the removed batched request iteration.
-    //   contextPoolSize  - promised "pre-initialized JS contexts", but the
-    //                      script context pool is grown lazily and never read a
-    //                      bound (issue #112). A user could set it 1..256 and
-    //                      change nothing.
-    //
-    // The 2026-08 sweep (#519) retired eleven more. Each was verified by
-    // searching the engine for a reader; where the constant behind the key kept
-    // its own callers, the constant stayed and only the knob went.
-    //
-    //   maxConnections       - no reader anywhere; the server enforces no such
-    //                          global limit.
-    //   tcpKeepAliveIdle     - curl_utils sets both CURL keep-alive options from
-    //   tcpKeepAliveInterval   constants and never consults the config.
-    //   statsInterval        - superseded by liveTickIntervalMs, which shares its
-    //                          default and its purpose and is the one that is
-    //                          read. Two rows, one mechanism.
-    //   maxJsonFieldSize     - json.cpp caps stored fields at the constant, so
-    //                          the "increase only if..." escape hatch the
-    //                          description offered did not exist.
-    //   sseConnectTimeout    - the dashboard's reconnect logic is renderer-side
-    //   sseMaxRetry            and never asked the engine for these; the sse::*
-    //   sseSendLastEventId     constants existed only to seed them.
-    //   dbTempStore          - all three are PRAGMAs the open callback applies
-    //   dbMmapSize             from constants. Wiring them buys nothing: the
-    //   dbWalAutocheckpoint    defaults are already the measured optimum (see
-    //                          docs/engine/benchmarks.md) and none has a user
-    //                          story. dbCacheSize, the one that does, is wired
-    //                          instead of retired.
-    for (const char* retired : { "requestBatchSize", "contextPoolSize",
-         "maxConnections", "tcpKeepAliveIdle", "tcpKeepAliveInterval",
-         "statsInterval", "maxJsonFieldSize", "sseConnectTimeout", "sseMaxRetry",
-         "sseSendLastEventId", "dbTempStore", "dbMmapSize", "dbWalAutocheckpoint" }) {
-        impl_->storage.remove_all<ConfigEntry> (
-        where (c (&ConfigEntry::key) == std::string (retired)));
-    }
-
-    // Get existing config entries (if any) to preserve user-modified values
-    auto existing = impl_->storage.get_all<ConfigEntry> ();
-    std::unordered_map<std::string, ConfigEntry> existing_map;
-    for (const auto& entry : existing) {
-        existing_map[entry.key] = entry;
-    }
-
-    auto now = std::chrono::duration_cast<std::chrono::milliseconds> (
-    std::chrono::system_clock::now ().time_since_epoch ())
-               .count ();
-
-    // Metadata markers, wrapped around the entry they apply to so the flag is
-    // read at the top of the seed rather than counted out as a trailing
-    // positional `true` after `now`.
-    //
-    // restart_required: the running engine keeps the old value until restarted.
-    // The app shows one chip from this and the Dock a pending signal; nothing
-    // states it in the label or the description any more.
-    //
-    // advanced: an internal with no everyday user story. Rendered collapsed
-    // under "Advanced" at the bottom of its category, not removed - a knob that
-    // is still live is still reachable.
-    auto restart_required = [] (ConfigEntry entry) {
-        entry.requires_restart = true;
-        return entry;
-    };
-    auto advanced = [] (ConfigEntry entry) {
-        entry.advanced = true;
-        return entry;
-    };
-    // keywords: the words a user arrives with that this entry's key, label and
-    // description never say - "ram" for a cache size, "deadline" for a timeout.
-    // Not a place to repeat what the entry already carries: search already
-    // matches all three fields, and a keyword that duplicates one only lifts
-    // the entry above better matches. A guard test enforces that.
-    auto keywords = [] (std::initializer_list<const char*> terms) {
-        // Serialized here rather than inside the returned wrapper: an
-        // initializer_list's backing array dies with the full-expression that
-        // built it, so a wrapper that held on to it would be reading freed
-        // storage the moment one is stored instead of called immediately.
-        std::string json =
-        nlohmann::json (std::vector<std::string> (terms.begin (), terms.end ())).dump ();
-        return [json = std::move (json)] (ConfigEntry entry) {
-            entry.keywords = json;
-            return entry;
-        };
-    };
-    // unit: what the value measures, for the entries that measure something.
-    // The app renders it as the input's suffix, which is the one place a unit
-    // belongs - so a description that also spells it out states it twice, and a
-    // guard test rejects the "in milliseconds" / "in bytes" clause for any
-    // entry that declares one. A count (workers, retained runs, stored steps)
-    // declares none: it measures nothing, and a suffix reading "items" is
-    // noise.
-    auto unit = [] (const char* symbol) {
-        return [symbol = std::string (symbol)] (ConfigEntry entry) {
-            entry.unit = symbol;
-            return entry;
-        };
-    };
-
-    // Helper lambda: Creates or updates a config entry
-    // - New entries get default values
-    // - Existing entries preserve user-modified values but get updated metadata
-    auto upsert_config = [&] (const ConfigEntry& new_entry) {
-        known_config_keys_.insert (new_entry.key);
-        auto it = existing_map.find (new_entry.key);
-        if (it != existing_map.end ()) {
-            // Preserve user's value but update metadata (description, label, etc.)
-            ConfigEntry updated = new_entry;
-            updated.value       = it->second.value;     // Keep user's value
-            updated.updated_at = it->second.updated_at; // Keep original timestamp
-            impl_->storage.replace (updated);
-        } else {
-            // New entry - use defaults
-            impl_->storage.replace (new_entry);
-        }
-    };
-
-    // =========================================================================
-    // CORE (general_engine)
-    // The engine's own machinery: how much work it runs in parallel, and how
-    // SQLite is tuned underneath it. Everything a *request* or a *run* is
-    // measured or bounded by lives in one of the categories below - #703 moved
-    // the request defaults to Network & connectivity and the ceilings to
-    // Limits, leaving the four settings that describe the engine itself.
-    // =========================================================================
-
-    // Not restart_required: `run_manager` reads this at the start of every run
-    // and builds that run's EventLoop from it, so a change is in force for the
-    // next run started. It carried the flag until #873, which told the user to
-    // restart for a setting the engine had already picked up.
-    upsert_config (keywords ({ "cores", "parallelism" }) (ConfigEntry{
-    "workers", std::to_string (std::thread::hardware_concurrency ()), "integer", "Worker Threads",
-    "Number of background worker threads. Higher values improve throughput on "
-    "multi-core systems but increase RAM usage. "
-    "Default equals CPU core count.",
-    "general_engine", std::to_string (std::thread::hardware_concurrency ()),
-    "1", "128", std::nullopt, now }));
-
-    upsert_config (restart_required (unit ("bytes") (keywords ({ "ram" }) (ConfigEntry{ "dbCacheSize",
-    std::to_string (vayu::core::constants::database::CACHE_SIZE_BYTES), "integer", "Database Cache Size",
-    "Memory SQLite keeps per connection for recently used database pages. A "
-    "larger cache spares repeated reads while a high-RPS run writes results "
-    "and "
-    "the dashboard queries them. 64 to 128 megabytes suits runs storing 10,000 "
-    "or more results a second.",
-    "general_engine", std::to_string (vayu::core::constants::database::CACHE_SIZE_BYTES),
-    "1048576",    // min: 1MB in bytes
-    "1073741824", // max: 1GB in bytes
-    std::nullopt, now }))));
-
-    upsert_config (restart_required (advanced (
-    unit ("ms") (keywords ({ "contention" }) (ConfigEntry{ "dbBusyTimeout",
-    std::to_string (vayu::core::constants::database::BUSY_TIMEOUT_MS), "integer", "Database Lock Wait Time",
-    "How long a thread waits for the database when another one is writing to "
-    "it. Result-storage threads compete for it during a high-concurrency run, "
-    "and a longer wait turns a 'database is locked' failure into a pause "
-    "instead. 10 to 30 seconds suits runs at 100 or more concurrent requests.",
-    "general_engine", std::to_string (vayu::core::constants::database::BUSY_TIMEOUT_MS),
-    "1000",  // min: 1 second
-    "60000", // max: 60 seconds
-    std::nullopt, now })))));
-
-    upsert_config (restart_required (keywords ({ "fsync", "pragma" }) (ConfigEntry{ "dbSynchronous",
-    std::to_string (vayu::core::constants::database::SYNCHRONOUS), "enum", "Data Safety Mode",
-    "How hard SQLite works to get results onto disk before reporting them "
-    "written. Off is the fastest and the default: the database stays "
-    "uncorrupted through a crash, but the last few results may be lost to a "
-    "power cut, which is an acceptable trade for test telemetry. Normal and "
-    "Full buy durability back at the cost of write throughput during a "
-    "high-RPS run.",
-    "general_engine", std::to_string (vayu::core::constants::database::SYNCHRONOUS),
-    std::nullopt, std::nullopt, db_synchronous_options_json (), now })));
-
-    // =========================================================================
-    // NETWORK & CONNECTIVITY (network_performance)
-    // The wire itself: what a new request starts with, how many transfers a
-    // worker keeps open, and how long a name resolution is reused. The OAuth
-    // renewal knobs that used to sit here were five of its eight entries and
-    // moved to Services (#703) - a user tuning token renewal does not arrive
-    // at "network tuning".
-    // =========================================================================
-
-    upsert_config (unit ("ms") (keywords ({ "deadline", "give up", "hang" }) (ConfigEntry{ "defaultTimeout",
-    std::to_string (vayu::core::constants::server::DEFAULT_TIMEOUT_MS), "integer", "Default Request Timeout",
-    "How long an HTTP request may run before it is abandoned, "
-    "when the request does not set a timeout of its own. Raise it for slow "
-    "endpoints; lower it to fail faster.",
-    "network_performance", std::to_string (vayu::core::constants::server::DEFAULT_TIMEOUT_MS),
-    "1000",   // min: 1 second
-    "300000", // max: 5 minutes
-    std::nullopt, now })));
-
-    upsert_config (
-    keywords ({ "h2", "alpn" }) (ConfigEntry{ "defaultHttpVersion",
-    vayu::to_string (vayu::DEFAULT_HTTP_VERSION), "enum", "Default HTTP Version",
-    "Protocol a newly created request starts with. Auto lets the server and "
-    "client negotiate (HTTP/2 where available, falling back to HTTP/1.1). "
-    "Changing this does not alter requests that already exist.",
-    "network_performance", vayu::to_string (vayu::DEFAULT_HTTP_VERSION),
-    std::nullopt, std::nullopt, http_version_options_json (), now }));
-
-    upsert_config (keywords ({ "concurrency", "parallel" }) (ConfigEntry{
-    "eventLoopMaxConcurrent", std::to_string (vayu::core::constants::event_loop::MAX_CONCURRENT),
-    "integer", "Max Concurrent Requests (Per Worker)",
-    "How many requests one worker keeps in flight at once. Higher values use "
-    "more file descriptors and push the target harder. Read when a load test "
-    "starts, so a change applies to the next run.",
-    "network_performance", std::to_string (vayu::core::constants::event_loop::MAX_CONCURRENT),
-    "1", "10000", std::nullopt, now }));
-
-    upsert_config (keywords ({ "socket", "pool" }) (ConfigEntry{
-    "eventLoopMaxPerHost", std::to_string (vayu::core::constants::event_loop::MAX_PER_HOST),
-    "integer", "Max Connections Per Host (Per Worker)",
-    "Concurrency limit for a specific target API host. Critical for respecting "
-    "target rate limits. "
-    "Lower values are gentler on the target; higher values maximize "
-    "throughput.",
-    "network_performance", std::to_string (vayu::core::constants::event_loop::MAX_PER_HOST),
-    "1", "1000", std::nullopt, now }));
-
-    upsert_config (
-    unit ("sec") (keywords ({ "ttl" }) (ConfigEntry{ "dnsCacheTimeout",
-    std::to_string (vayu::core::constants::event_loop::DNS_CACHE_TIMEOUT_SECONDS), "integer", "DNS Cache Timeout",
-    "How long a resolved hostname is reused before it is looked up "
-    "again. 0 forces a fresh lookup on every request; 60 to 300 seconds suits "
-    "a "
-    "stable endpoint.",
-    "network_performance", std::to_string (vayu::core::constants::event_loop::DNS_CACHE_TIMEOUT_SECONDS),
-    "0",    // Disable cache
-    "3600", // 1 hour
-    std::nullopt, now })));
-
-    // Proxy (issue #705). Three entries, read together by
-    // `resolve_transport_policy` at the point of use, so a change applies to
-    // the next transfer rather than the next restart. The keywords are the
-    // words someone arrives with when nothing works and they suspect the
-    // network - none of which the labels say.
-    upsert_config (keywords ({ "corporate", "firewall", "mitm", "zscaler", "vpn" }) (
-    ConfigEntry{ "proxyMode",
-    vayu::http::to_string (vayu::http::TransportPolicy{}.proxy_mode), "enum", "Proxy",
-    "How outbound requests reach the network. From environment uses the "
-    "http_proxy and https_proxy variables the engine was started with, which "
-    "is what a terminal-launched engine already picks up and a desktop launch "
-    "usually has none of. From system uses the proxy this computer is "
-    "configured with, which the app resolves and shows below. Manual routes "
-    "everything through the proxy URL below. None sends direct, ignoring those "
-    "variables too.",
-    "network_performance", vayu::http::to_string (vayu::http::TransportPolicy{}.proxy_mode),
-    std::nullopt, std::nullopt, proxy_mode_options_json (), now }));
-
-    upsert_config (keywords ({ "corporate", "firewall", "mitm" }) (
-    ConfigEntry{ "proxyUrl", "", "string", "Proxy URL",
-    "The proxy to route through when Proxy is set to Manual, written the way "
-    "curl takes it: scheme://user:password@host:port. The scheme selects the "
-    "kind - http, https, socks4, socks4a, socks5, socks5h - and credentials in "
-    "the URL are sent as basic proxy authentication.",
-    "network_performance", "", std::nullopt, std::nullopt, std::nullopt, now }));
-
-    // The resolved system proxy (issue #708). Written by the app's main
-    // process, which is the only part of Vayu that can ask the operating system
-    // - Chromium resolves it, libcurl sees none of it - and read by the engine
-    // only under `proxyMode: system`.
-    //
-    // A visible row rather than a hidden side channel: a proxy the user cannot
-    // see is the failure this epic exists to end, and "system" that silently
-    // resolved to nothing has to be readable as such. It is stored where the
-    // user can edit it, and an edit lasts until the next resolution overwrites
-    // it - said in the description rather than enforced, because a read-only
-    // entry type would be a new concept in the config table for one row.
-    upsert_config (keywords ({ "wpad", "autoconfig", "automatic" }) (
-    ConfigEntry{ "proxySystemUrl", "", "string", "System Proxy (resolved)",
-    "The proxy this computer is configured with, as the app resolved it when "
-    "it started and whenever the network changed. Read only when Proxy is set "
-    "to From system; empty means nothing resolved - a direct configuration, or "
-    "an engine running with no app to ask - and requests then fall back to the "
-    "http_proxy and https_proxy variables. A PAC script is resolved once, "
-    "against a sample URL, and the one answer applies to every request: a "
-    "configuration that returns different proxies for different URLs needs "
-    "Manual instead. Typing a value here works until the next resolution "
-    "replaces it.",
-    "network_performance", "", std::nullopt, std::nullopt, std::nullopt, now }));
-
-    upsert_config (keywords ({ "exclude", "whitelist", "intranet" }) (
-    ConfigEntry{ "proxyBypass", "", "string", "Proxy Bypass List",
-    "Hosts that skip the proxy, comma-separated. A leading dot matches a "
-    "domain and everything under it (.internal.example.com), and a single * "
-    "bypasses the proxy for every host. Under Manual this list is the whole "
-    "rule, so an empty one means nothing is exempt and any no_proxy the engine "
-    "was started with is ignored. Under From environment it overrides that "
-    "variable when set, and defers to it when empty. Under From system it "
-    "follows whichever of those two that mode resolved to.",
-    "network_performance", "", std::nullopt, std::nullopt, std::nullopt, now }));
-
-    // What the engine adds to a request nobody wrote it into (issue #1229).
-    // Four entries, read together by `resolve_default_header_policy` at the top
-    // of a request or a run, so a change applies to the next send rather than
-    // the next restart - the shape the proxy entries above use, and for the
-    // same reason. What each of them adds is shown in the Headers tab as a row
-    // the request can switch off; none of them is stored in the request.
-    upsert_config (keywords ({ "gzip", "brotli", "deflate", "zstd" }) (
-    ConfigEntry{ "negotiateCompression", "true", "boolean", "Negotiate Compressed Responses",
-    "Asks servers for a compressed response - the Accept-Encoding header a "
-    "browser, Postman and curl --compressed all send - and decodes it before "
-    "you see it. Response sizes and times then describe what production does. "
-    "Off sends no Accept-Encoding at all, so a server that would have "
-    "compressed answers with the identity bytes instead. Applies to a Send, a "
-    "collection run and a script's own request; a load run has its own setting "
-    "below. A request that carries its own Accept-Encoding header is left "
-    "alone, and its response is not decoded.",
-    "network_performance", "true", std::nullopt, std::nullopt, std::nullopt, now }));
-
-    upsert_config (keywords ({ "gzip", "brotli", "zstd", "throughput" }) (
-    ConfigEntry{ "loadNegotiateCompression", "true", "boolean", "Negotiate Compressed Responses (Load Tests)",
-    "The same decision for a load run, kept separate because compression is "
-    "part of what a load test measures: with it on, the bytes counted are what "
-    "the server sent and the sizes reported are what Vayu decoded; with it "
-    "off, both are the identity response. Turn it off to measure a server's "
-    "uncompressed ceiling, on to measure what its clients actually get.",
-    "network_performance", "true", std::nullopt, std::nullopt, std::nullopt, now }));
-
-    upsert_config (keywords ({ "tracing", "x-request-id", "debugging" }) (
-    ConfigEntry{ "correlationIdEnabled", "false", "boolean", "Send a Correlation Id",
-    "Adds a header carrying a fresh identifier to every request, so one send "
-    "or one iteration of a load run can be found in a server's logs. Off by "
-    "default: it is a header your server did not ask for, and a gateway that "
-    "gives the name its own meaning would see one it did not expect. A request "
-    "that carries the header itself is left alone.",
-    "network_performance", "false", std::nullopt, std::nullopt, std::nullopt, now }));
-
-    upsert_config (
-    keywords ({ "tracing", "distributed" }) (ConfigEntry{ "correlationIdHeader",
-    std::string (vayu::http::DEFAULT_CORRELATION_HEADER), "string", "Correlation Id Header",
-    "Which header the correlation id goes out under, when it is switched on. "
-    "The default is namespaced to Vayu so it collides with nothing; set it to "
-    "the name your own infrastructure reads - X-Request-ID and "
-    "X-Correlation-ID are the common ones - if you want Vayu's id to land "
-    "there instead.",
-    "network_performance", std::string (vayu::http::DEFAULT_CORRELATION_HEADER),
-    std::nullopt, std::nullopt, std::nullopt, now }));
-
-    // Custom trust anchors (issue #706). `text` rather than `string` because
-    // what goes in it is a pasted PEM block: the single-line input every other
-    // string entry renders cannot show one, let alone several.
-    //
-    // Content, not a path - a path breaks the moment the file moves and cannot
-    // be shown back in Settings. The engine materializes the bundle beside the
-    // database, extending the platform's own anchors rather than replacing
-    // them; see `resolve_transport_policy`.
-    upsert_config (keywords ({ "firewall", "mitm", "zscaler", "ssl" }) (
-    ConfigEntry{ "customCaCertificates", "", "text", "Custom CA Certificates",
-    "Certificate authorities to trust in addition to the ones this platform "
-    "already trusts, pasted as PEM text (one or more -----BEGIN "
-    "CERTIFICATE----- blocks). This is what makes a corporate TLS-inspecting "
-    "proxy or an internal self-signed authority verifiable everywhere the "
-    "engine sends: requests, load runs, streams, OAuth token fetches and spec "
-    "imports alike. Paste certificates only - never a private key.",
-    "network_performance", "", std::nullopt, std::nullopt, std::nullopt, now }));
-
-    // =========================================================================
-    // SERVICES (services)
-    // The long-lived surfaces: streaming requests, the webhook inboxes and mock
-    // servers the Dock calls Services, and the OAuth issuers behind them.
-    // Before #586 these were filed under Observability, which is the drawer's
-    // word for none of them - a user managing a service found no matching word
-    // in the tree.
-    // =========================================================================
-
-    // SSE requests (issue #573). All six - these five and `sseMaxStoredEvents`,
-    // which #703 moved to Data & retention as the per-run storage budget it is -
-    // are read once when a stream starts, so a change applies to the next
-    // stream - no restart - and one run's events are all bounded by a single
-    // rule rather than by whatever the setting happened to be at each arrival.
-    upsert_config (keywords ({ "eventsource", "server-sent", "event stream" }) (
-    ConfigEntry{ "sseMaxRetainedEvents",
-    std::to_string (vayu::core::constants::sse::MAX_RETAINED_EVENTS), "integer", "Stream Events Retained",
-    "How many of a streaming request's events are held in memory for the "
-    "Events "
-    "timeline. Older ones are dropped as new ones arrive, so this is how far "
-    "back a long stream can be scrolled while it is running; the completed run "
-    "stores its own, separate list. Each event costs roughly its own size in "
-    "memory.",
-    "services", std::to_string (vayu::core::constants::sse::MAX_RETAINED_EVENTS),
-    std::to_string (vayu::core::constants::sse::MIN_RETAINED_EVENTS),
-    std::to_string (vayu::core::constants::sse::RETAINED_EVENTS_CEILING),
-    std::nullopt, now }));
-
-    upsert_config (unit ("bytes") (
-    keywords ({ "eventsource", "server-sent", "event stream" }) (ConfigEntry{ "sseMaxEventBytes",
-    std::to_string (vayu::core::constants::sse::MAX_EVENT_BYTES), "integer", "Stream Event Size Limit",
-    "How much of a single streamed event is kept. A larger event is held as a "
-    "prefix and flagged as truncated, never silently cut - the event reports "
-    "the size as received either way. This is also what bounds the parser "
-    "itself, so a server that sends without a line break cannot exhaust "
-    "memory.",
-    "services", std::to_string (vayu::core::constants::sse::MAX_EVENT_BYTES),
-    std::to_string (vayu::core::constants::sse::MIN_EVENT_BYTES),
-    std::to_string (vayu::core::constants::sse::EVENT_BYTES_CEILING), std::nullopt, now })));
-
-    upsert_config (unit ("ms") (keywords ({ "eventsource", "server-sent",
-    "event stream" }) (ConfigEntry{ "sseMaxStreamDurationMs",
-    std::to_string (vayu::core::constants::sse::MAX_STREAM_DURATION_MS), "integer", "Stream Duration Limit",
-    "How long a streaming request may run before the engine ends it and says "
-    "so. A stream has no content length and no promise to end, so this is the "
-    "backstop that keeps one from holding a worker forever. A request may ask "
-    "for a shorter limit of its own.",
-    "services", std::to_string (vayu::core::constants::sse::MAX_STREAM_DURATION_MS),
-    std::to_string (vayu::core::constants::sse::MIN_STREAM_DURATION_MS),
-    std::to_string (vayu::core::constants::sse::STREAM_DURATION_MS_CEILING),
-    std::nullopt, now })));
-
-    upsert_config (keywords ({ "eventsource", "server-sent", "event stream" }) (
-    ConfigEntry{ "sseMaxStreamEvents",
-    std::to_string (vayu::core::constants::sse::MAX_STREAM_EVENTS), "integer", "Stream Event Limit",
-    "How many events a streaming request may receive before the engine ends it "
-    "and says so. The count backstop beside the time one, for a stream that "
-    "talks fast rather than long. A request may ask for a lower limit of its "
-    "own.",
-    "services", std::to_string (vayu::core::constants::sse::MAX_STREAM_EVENTS),
-    std::to_string (vayu::core::constants::sse::MIN_STREAM_EVENTS),
-    std::to_string (vayu::core::constants::sse::STREAM_EVENTS_CEILING), std::nullopt, now }));
-
-    upsert_config (advanced (unit ("ms") (
-    keywords ({ "eventsource", "server-sent", "event stream" }) (ConfigEntry{ "sseIdleTimeoutMs",
-    std::to_string (vayu::core::constants::sse::IDLE_TIMEOUT_MS), "integer", "Stream Idle Timeout",
-    "How long a stream may deliver nothing before the engine ends it. This is "
-    "the one deadline a stream gets - a whole-transfer timeout would kill a "
-    "healthy stream mid-flight - so it must be longer than the quietest gap "
-    "the endpoint you are watching leaves between events. Enforced at whole-"
-    "second resolution.",
-    "services", std::to_string (vayu::core::constants::sse::IDLE_TIMEOUT_MS),
-    std::to_string (vayu::core::constants::sse::MIN_IDLE_TIMEOUT_MS),
-    std::to_string (vayu::core::constants::sse::IDLE_TIMEOUT_MS_CEILING), std::nullopt, now }))));
-
-    // Webhook inbox. All three are read once when an inbox starts, so a change
-    // applies to the next inbox started - no restart. The running listener keeps
-    // what it was started with, which is what makes one inbox's captures a set
-    // truncated and retained by a single rule rather than by whatever the
-    // setting happened to be at each arrival.
-    upsert_config (unit ("bytes") (ConfigEntry{ "inboxMaxBodyBytes",
-    std::to_string (vayu::core::constants::inbox::MAX_BODY_BYTES), "integer", "Inbox Capture Body Limit",
-    "How much of an inbound webhook body an inbox stores. A larger "
-    "payload is kept as a prefix and flagged as truncated, never silently cut "
-    "- "
-    "the capture reports the size as received either way. Raise it for a "
-    "provider that posts large documents; the transport still refuses anything "
-    "over 8 MB outright, which is not a webhook.",
-    "services", std::to_string (vayu::core::constants::inbox::MAX_BODY_BYTES),
-    std::to_string (vayu::core::constants::inbox::MIN_BODY_BYTES),
-    std::to_string (vayu::core::constants::inbox::MAX_PAYLOAD_BYTES), std::nullopt, now }));
-
-    upsert_config (ConfigEntry{ "inboxMaxCaptures",
-    std::to_string (vayu::core::constants::inbox::MAX_CAPTURES), "integer", "Inbox Captures Retained",
-    "How many requests one inbox keeps before the oldest are dropped. Also the "
-    "most a single page of the capture list may ask for. Raise it to keep a "
-    "long webhook session whole; every capture is a stored row, so this and "
-    "the body limit above together bound what an inbox costs on disk.",
-    "services", std::to_string (vayu::core::constants::inbox::MAX_CAPTURES),
-    std::to_string (vayu::core::constants::inbox::MIN_CAPTURES),
-    std::to_string (vayu::core::constants::inbox::CAPTURES_CEILING), std::nullopt, now });
-
-    upsert_config (
-    advanced (unit ("ms") (ConfigEntry{ "inboxLivePollIntervalMs",
-    std::to_string (vayu::core::constants::inbox::LIVE_POLL_INTERVAL_MS), "integer", "Inbox Live Poll Interval",
-    "How often a watched inbox checks for newly arrived "
-    "captures. This is the delay between a webhook landing and its row "
-    "appearing. Lower costs a few more wakeups per second on the one thread "
-    "holding that stream and nothing on the capture path itself.",
-    "services", std::to_string (vayu::core::constants::inbox::LIVE_POLL_INTERVAL_MS),
-    std::to_string (vayu::core::constants::inbox::MIN_LIVE_POLL_INTERVAL_MS),
-    std::to_string (vayu::core::constants::inbox::MAX_LIVE_POLL_INTERVAL_MS),
-    std::nullopt, now })));
-
-    // Mid-run OAuth 2.0 refresh. A load run renews a header-placed access token
-    // before it expires, so a run longer than its token does not turn into a
-    // 401 storm. All five are read once when the run arms its watchdog, so a
-    // change applies to the next run started - no restart. They sat in Network
-    // & connectivity until #703, where they were five of its eight entries; a
-    // user tuning token renewal arrives at Services, which is where the Dock
-    // files the OAuth issuer itself.
-    upsert_config (unit ("ms") (ConfigEntry{ "oauth2RefreshLeadMs",
-    std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_LEAD_MS),
-    "integer", "OAuth 2.0 Refresh Lead Time",
-    "How far ahead of an access token's expiry a running load "
-    "test renews it. Wider than the 45-second skew the token cache already "
-    "applies, so the new credential is published while the old one is still "
-    "accepted and no request falls in the gap. Raise it for a provider that is "
-    "slow to issue tokens.",
-    "services", std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_LEAD_MS),
-    "1000",    // 1 second
-    "3600000", // 1 hour
-    std::nullopt, now }));
-
-    upsert_config (
-    advanced (unit ("ms") (ConfigEntry{ "oauth2RefreshMinIntervalMs",
-    std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_MIN_INTERVAL_MS),
-    "integer", "Min OAuth 2.0 Refresh Interval",
-    "Floor on the wait between two mid-run renewals. A token "
-    "whose whole lifetime is shorter than the lead time above is always inside "
-    "its refresh window, so without this floor a run would re-acquire in a "
-    "tight "
-    "loop and hammer the token endpoint. Lower it only for a provider that "
-    "issues very short-lived tokens.",
-    "services", std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_MIN_INTERVAL_MS),
-    "100",     // 0.1 second - a floor, never 0: that is the tight loop
-    "3600000", // 1 hour
-    std::nullopt, now })));
-
-    upsert_config (advanced (unit ("ms") (ConfigEntry{ "oauth2RefreshRetryMs",
-    std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_RETRY_MS), "integer", "OAuth 2.0 Refresh Retry Delay",
-    "First wait after a mid-run renewal is refused, doubled "
-    "per consecutive failure up to the ceiling below. The run keeps sending "
-    "the "
-    "credential it already has - a failed renewal is reported in the run's "
-    "report, never fatal - so this is about recovering from a token endpoint "
-    "that blipped.",
-    "services", std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_RETRY_MS),
-    "250", "600000", std::nullopt, now })));
-
-    upsert_config (
-    advanced (unit ("ms") (ConfigEntry{ "oauth2RefreshRetryMaxMs",
-    std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_RETRY_MAX_MS),
-    "integer", "Max OAuth 2.0 Refresh Retry Delay",
-    "Ceiling on that backoff, so a token endpoint that is "
-    "down for an hour costs the run a bounded number of attempts rather than "
-    "one every few seconds.",
-    "services", std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_RETRY_MAX_MS),
-    "1000", "3600000", std::nullopt, now })));
-
-    upsert_config (
-    advanced (unit ("ms") (ConfigEntry{ "oauth2RefreshPollIntervalMs",
-    std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_POLL_INTERVAL_MS),
-    "integer", "OAuth 2.0 Refresh Poll Interval",
-    "How often the renewal watchdog wakes while it waits, to "
-    "notice that the run has ended. A finished run joins that thread before it "
-    "writes its report, so this bounds how long the run's last moments take. "
-    "Lower costs a few more wakeups per second on one sleeping thread and "
-    "nothing on the request path.",
-    "services", std::to_string (vayu::core::constants::server::OAUTH2_REFRESH_POLL_INTERVAL_MS),
-    "10",   // 10ms - below this the wakeups outweigh what they save
-    "5000", // 5s - past this a finished run visibly waits on the join
-    std::nullopt, now })));
-
-    // =========================================================================
-    // OBSERVABILITY (observability)
-    // What a run measures and what the dashboard's live charts are fed: the
-    // server-vitals monitor, the live-metrics topic, and the per-phase timing
-    // histograms. What a run *stores* is Data & retention below - the two were
-    // one category holding 24 of 48 entries until #586.
-    // =========================================================================
-
-    // Server-vitals monitor. All three are read per run - a change applies to
-    // the next run started, no restart. The interval *bounds* (250-60000ms) are
-    // deliberately not settings: they exist to stop a cadence that measures the
-    // scraper rather than the target.
-    upsert_config (
-    unit ("ms") (keywords ({ "prometheus" }) (ConfigEntry{ "monitorIntervalMs",
-    std::to_string (vayu::core::constants::monitor::DEFAULT_INTERVAL_MS), "integer", "Server Monitoring Scrape Interval",
-    "How often a load test scrapes the metrics endpoint it "
-    "was pointed at, when the run does not set its own interval. Each scrape "
-    "is "
-    "one request on the run's monitor thread, so it never delays the run's own "
-    "metrics - but a cadence faster than the target's own collection interval "
-    "only re-reads the same numbers. Raise it for an endpoint that is "
-    "expensive "
-    "to render.",
-    "observability", std::to_string (vayu::core::constants::monitor::DEFAULT_INTERVAL_MS),
-    std::to_string (vayu::core::constants::monitor::MIN_INTERVAL_MS),
-    std::to_string (vayu::core::constants::monitor::MAX_INTERVAL_MS), std::nullopt, now })));
-
-    upsert_config (keywords ({ "prometheus" }) (ConfigEntry{ "monitorMaxSeries",
-    std::to_string (vayu::core::constants::monitor::MAX_SERIES), "integer", "Server Monitoring Metric Limit",
-    "How many metric names one run may chart from its monitored endpoint. Each "
-    "is a line on a single overlay and a name matched against every line of "
-    "the "
-    "exposition body, so the ceiling is about a readable chart rather than a "
-    "hard cost. The chart has four distinct colours and repeats them past "
-    "that.",
-    "observability", std::to_string (vayu::core::constants::monitor::MAX_SERIES),
-    "1", "64", std::nullopt, now }));
-
-    upsert_config (advanced (unit ("ms") (keywords ({ "prometheus" }) (ConfigEntry{ "monitorScrapeTimeoutMs",
-    std::to_string (vayu::core::constants::monitor::DEFAULT_SCRAPE_TIMEOUT_MS), "integer", "Server Monitoring Scrape Timeout",
-    "How long one scrape of the metrics endpoint may take "
-    "before it counts as a gap in the series. 0 derives the budget from the "
-    "scrape interval - three quarters of it - which is what most endpoints "
-    "want; set it explicitly for an exposition slow enough to fail every "
-    "scrape "
-    "at that budget, where the only other way out is a slower cadence that "
-    "also "
-    "thins the data. A value longer than the interval a run scrapes at is "
-    "shortened to it, because a scrape that outlives its own cadence puts the "
-    "loop behind itself.",
-    "observability", std::to_string (vayu::core::constants::monitor::DEFAULT_SCRAPE_TIMEOUT_MS),
-    "0", std::to_string (vayu::core::constants::monitor::MAX_INTERVAL_MS),
-    std::nullopt, now }))));
-
-    upsert_config (unit ("ms") (keywords ({ "refresh rate", "sse" }) (ConfigEntry{ "liveTickIntervalMs",
-    std::to_string (vayu::core::constants::server::STATS_INTERVAL_MS), "integer", "Live Metrics Tick Interval",
-    "How often the engine emits a live-metrics tick into the "
-    "in-memory replay topic during a run. A lower value gives smoother live "
-    "charts for slightly more CPU; the 1-second ceiling exists because slower "
-    "ticks defeat live smoothness. The historical 1 Hz database sampling is "
-    "unaffected.",
-    "observability", std::to_string (vayu::core::constants::server::STATS_INTERVAL_MS),
-    "10", "1000", std::nullopt, now })));
-
-    upsert_config (
-    unit ("ms") (keywords ({ "time range" }) (ConfigEntry{ "liveReplayWindowMs",
-    std::to_string (vayu::core::constants::server::DEFAULT_LIVE_REPLAY_WINDOW_MS), "integer", "Live Chart Window",
-    "How much recent live-metrics history to keep: the span the dashboard's "
-    "live "
-    "charts show, and the span the engine holds in memory per run so the "
-    "dashboard can rebuild those charts when it attaches - or re-attaches - "
-    "mid-run. One setting drives both, so they cannot disagree. Its editor is "
-    "Settings > Dashboard > Chart window, which is where the effect is "
-    "visible; this list deliberately does not offer a second one. Expressed "
-    "as elapsed time "
-    "rather than a tick count, so it survives a change to the tick interval. 0 "
-    "means the full run (no time limit). Live Metrics Tick Ceiling is the "
-    "memory backstop "
-    "either way, so a fast tick interval reaches that ceiling before a long "
-    "window does.",
-    "observability", std::to_string (vayu::core::constants::server::DEFAULT_LIVE_REPLAY_WINDOW_MS),
-    "0", "3600000", std::nullopt, now })));
-
-    upsert_config (advanced (ConfigEntry{ "liveMaxRetainedTicks",
-    std::to_string (vayu::core::constants::server::DEFAULT_MAX_LIVE_TICKS), "integer", "Live Metrics Tick Ceiling",
-    "Hard ceiling on live-metrics data points held in memory per run, on both "
-    "sides - the engine's replay ring and the dashboard's chart history. It is "
-    "a memory bound rather than a rendering one, since the charts bucket "
-    "points "
-    "before plotting, and it binds only when the chart window divided by the "
-    "tick interval exceeds it - a long window at a fast tick interval, which "
-    "stock settings never reach. Raise it if a long window is being cut short; "
-    "each point costs roughly 1 KB.",
-    "observability", std::to_string (vayu::core::constants::server::DEFAULT_MAX_LIVE_TICKS),
-    "1000", "500000", std::nullopt, now }));
-
-    upsert_config (unit ("ms") (ConfigEntry{ "liveRetentionMs", "60000", "integer", "Live Metrics Retention",
-    "How long a finished run's in-memory live-metrics topic "
-    "is kept so the dashboard can still attach and replay it. After this "
-    "window "
-    "the run is evicted and the dashboard falls back to the stored report. 0 "
-    "disables retention, so that fallback is immediate.",
-    "observability", "60000", "0", "600000", std::nullopt, now }));
-
-    upsert_config (restart_required (keywords ({ "verbosity", "logging" }) (
-    ConfigEntry{ "logLevel", vayu::core::constants::logging::DEFAULT_LEVEL, "enum", "Engine Log Level",
-    "The lowest severity the engine writes to its log file. Debug records "
-    "everything, which is what a bug report wants and what fills a disk "
-    "fastest; "
-    "Warning and Error keep a long-running install quiet. The console is "
-    "separate - it follows the daemon's -v flag, so raising this does not "
-    "silence a terminal you started the engine in.",
-    "observability", vayu::core::constants::logging::DEFAULT_LEVEL,
-    std::nullopt, std::nullopt, log_level_options_json (), now })));
-
-    upsert_config (advanced (restart_required (unit ("bytes") (
-    keywords ({ "rotation", "retention" }) (ConfigEntry{ "maxLogFileBytes",
-    std::to_string (vayu::core::constants::logging::DEFAULT_MAX_FILE_BYTES), "integer", "Max Log File Size",
-    "How large one log file may grow before it is rotated once to a '.1' "
-    "beside it and writing continues in a fresh one. A start already gets its "
-    "own file, and the newest " +
-    std::to_string (vayu::core::constants::logging::RETAINED_FILES) +
-    " of those are what the log directory keeps, so this bounds the one case "
-    "that naming cannot: a single run chatty enough to fill the disk by "
-    "itself. 0 removes the bound.",
-    "observability", std::to_string (vayu::core::constants::logging::DEFAULT_MAX_FILE_BYTES),
-    "0", "1073741824", std::nullopt, now })))));
-
-    upsert_config (keywords ({ "ttfb", "timings" }) (ConfigEntry{
-    "phaseHistograms", vayu::core::constants::metrics_collector::DEFAULT_PHASE_HISTOGRAMS ? "true" : "false",
-    "boolean", "Per-Phase Latency Histograms",
-    "Records DNS, connect, TLS, first-byte and download times for every "
-    "load-test completion, so the report gives each phase real percentiles "
-    "instead of an average over the few exchanges it stores traces for. This "
-    "is "
-    "what answers whether a slow p99 came from the server or from connection "
-    "setup. It costs five histogram writes per completion; turn it off only if "
-    "a run at your throughput ceiling measurably suffers.",
-    "observability", vayu::core::constants::metrics_collector::DEFAULT_PHASE_HISTOGRAMS ? "true" : "false",
-    std::nullopt, std::nullopt, std::nullopt, now }));
-
-    // =========================================================================
-    // DATA & RETENTION (data_retention)
-    // Every per-run storage budget, in one place: how much of a body is kept,
-    // how many records are kept, and how long a finished run survives. The
-    // words a user arrives with here are "why is my response body cut off" and
-    // "keep fewer runs". #703 completed the shelf - the per-step and the
-    // per-stream retention budgets were filed under Core and Services.
-    // =========================================================================
-
-    upsert_config (ConfigEntry{ "maxStoredErrors",
-    std::to_string (vayu::core::constants::metrics_collector::DEFAULT_MAX_ERRORS),
-    "integer", "Stored Error Records Per Run",
-    "How many individual error records a run keeps for its report. The error "
-    "total, the failed-request count, the error rate and the status-code "
-    "breakdown are always exact - this bounds only the per-error detail behind "
-    "the report's 'By Error Type' breakdown, which on a run with more errors "
-    "than this covers the first N and will not sum to the total beside it. 0 "
-    "means unlimited, which against a fully refusing target grows for the life "
-    "of the run.",
-    "data_retention", std::to_string (vayu::core::constants::metrics_collector::DEFAULT_MAX_ERRORS),
-    "0", "10000000", std::nullopt, now });
-
-    upsert_config (unit ("bytes") (ConfigEntry{ "maxTraceBodyBytes",
-    std::to_string (vayu::core::constants::json::MAX_TRACE_BODY_BYTES), "integer", "Max Stored Trace Body Size",
-    "Largest request or response body kept in a design run's stored "
-    "trace. A larger body is truncated in the database - the response viewer "
-    "says so, and re-sending fetches the full body - so one huge response does "
-    "not bloat storage forever.",
-    "data_retention", std::to_string (vayu::core::constants::json::MAX_TRACE_BODY_BYTES),
-    "1024",      // 1KB
-    "104857600", // 100MB
-    std::nullopt, now }));
-
-    upsert_config (unit ("bytes") (ConfigEntry{ "maxSampleBodyBytes",
-    std::to_string (vayu::core::constants::metrics_collector::DEFAULT_MAX_SAMPLE_BODY_BYTES),
-    "integer", "Max Captured Sample Body",
-    "Largest response body kept for a single captured load-run "
-    "sample. Deliberately far smaller than the design-run trace limit, because "
-    "a load run captures tens of exchanges nobody asked for individually. A "
-    "larger body is stored truncated and marked as such.",
-    "data_retention", std::to_string (vayu::core::constants::metrics_collector::DEFAULT_MAX_SAMPLE_BODY_BYTES),
-    "0",         // 0 disables body capture while keeping headers and metadata
-    "104857600", // 100MB
-    std::nullopt, now }));
-
-    upsert_config (unit ("bytes") (ConfigEntry{ "maxSampleBytes",
-    std::to_string (vayu::core::constants::metrics_collector::DEFAULT_MAX_SAMPLE_BYTES),
-    "integer", "Load-Run Capture Budget",
-    "How much captured response-body data one load run may store. Once spent, "
-    "samples keep their headers and metadata and only their bodies are "
-    "dropped, and the report says how many. Captured data is stored verbatim, "
-    "can contain credentials, and is deleted with the run.",
-    "data_retention", std::to_string (vayu::core::constants::metrics_collector::DEFAULT_MAX_SAMPLE_BYTES),
-    "0",          // 0 disables body capture while keeping headers and metadata
-    "1073741824", // 1GB
-    std::nullopt, now }));
-
-    upsert_config (unit ("bytes") (ConfigEntry{ "maxResponseSampleBytes",
-    std::to_string (vayu::core::constants::metrics_collector::DEFAULT_MAX_RESPONSE_SAMPLE_BYTES),
-    "integer", "Load-Run Validation Sample Budget",
-    "How much response-body data one load run may hold for its post-run test "
-    "scripts and schema checks. These bodies are kept whole - a truncated one "
-    "would fail a check the target passed - so past the budget whole samples "
-    "are dropped instead, and the report counts them. Lower it for a target "
-    "with large responses; raise it to validate more of them.",
-    "data_retention", std::to_string (vayu::core::constants::metrics_collector::DEFAULT_MAX_RESPONSE_SAMPLE_BYTES),
-    "0",          // 0 retains no sample that has a body
-    "1073741824", // 1GB
-    std::nullopt, now }));
-
-    upsert_config (ConfigEntry{ "maxScenarioStoredSteps",
-    std::to_string (vayu::core::constants::scenario::MAX_STORED_STEPS),
-    "integer", "Max Stored Scenario Steps",
-    "How many per-step results one collection run keeps, which bounds what a "
-    "long run costs the dashboard to load. Steps that failed, errored or were "
-    "skipped are kept first and successes fill the rest, so raising this buys "
-    "more successful steps to look at and never a failure that was hidden - "
-    "what was thinned is reported in the run summary. 0 stores every step.",
-    "data_retention", std::to_string (vayu::core::constants::scenario::MAX_STORED_STEPS),
-    "0", "1000000", std::nullopt, now });
-
-    upsert_config (keywords ({ "eventsource", "server-sent", "event stream" }) (
-    ConfigEntry{ "sseMaxStoredEvents",
-    std::to_string (vayu::core::constants::sse::MAX_STORED_EVENTS), "integer", "Stream Events Stored Per Run",
-    "How many events a finished streaming run keeps on disk, so reopening it "
-    "from History shows the timeline again. A run that received more says so - "
-    "the stored list is marked truncated and carries the true total. 0 keeps "
-    "the count and no events.",
-    "data_retention", std::to_string (vayu::core::constants::sse::MAX_STORED_EVENTS),
-    "0", std::to_string (vayu::core::constants::sse::STORED_EVENTS_CEILING),
-    std::nullopt, now }));
-
-    upsert_config (keywords ({ "cleanup" }) (ConfigEntry{ "maxRunsRetained",
-    std::to_string (vayu::core::constants::database::MAX_RUNS_RETAINED), "integer", "Max Runs Retained",
-    "Keep at most this many most-recent runs; older runs, with their metrics "
-    "and results, are pruned at startup and after each run finishes. A higher "
-    "value - or 0 for unlimited - keeps more history but grows the database "
-    "file on disk and slows down loading the run history. In-progress runs are "
-    "never pruned.",
-    "data_retention", std::to_string (vayu::core::constants::database::MAX_RUNS_RETAINED),
-    "0", "100000", std::nullopt, now }));
-
-    upsert_config (
-    unit ("days") (keywords ({ "cleanup" }) (ConfigEntry{ "runRetentionDays",
-    std::to_string (vayu::core::constants::database::RUN_RETENTION_DAYS), "integer", "Run Retention",
-    "Delete runs older than this age, with their metrics and results, at "
-    "startup and after each run finishes. A higher value - or 0 to keep runs "
-    "forever - retains more history at the cost of a larger database file on "
-    "disk. In-progress runs are never pruned.",
-    "data_retention", std::to_string (vayu::core::constants::database::RUN_RETENTION_DAYS),
-    "0", "3650", std::nullopt, now })));
-
-    upsert_config (unit ("days") (keywords (
-    { "recycle bin", "recover", "cleanup" }) (ConfigEntry{ "trashRetentionDays",
-    std::to_string (vayu::core::constants::database::TRASH_RETENTION_DAYS), "integer", "Trash Retention",
-    "Deleted collections and requests are kept in the Trash this long before "
-    "they are destroyed for good; the sweep runs when Vayu starts. Until then "
-    "they can be restored exactly as they were. A higher value - or 0 to keep "
-    "them forever - leaves more to undo at the cost of a larger database file "
-    "on disk.",
-    "data_retention", std::to_string (vayu::core::constants::database::TRASH_RETENTION_DAYS),
-    "0", "3650", std::nullopt, now })));
-
-    upsert_config (
-    keywords ({ "restore", "cleanup" }) (ConfigEntry{ "maxBackupsRetained",
-    std::to_string (vayu::core::constants::database::MAX_BACKUPS_RETAINED), "integer", "Max Backups Retained",
-    "Keep at most this many workspace snapshots in the backups folder beside "
-    "the database; older ones are removed after each new backup. Each snapshot "
-    "is a compacted copy of the whole workspace - collections, environments, "
-    "secrets and run history - so a higher value, or 0 for unlimited, costs "
-    "disk. Only files Vayu wrote are ever removed; a copy you put there "
-    "yourself is left alone.",
-    "data_retention", std::to_string (vayu::core::constants::database::MAX_BACKUPS_RETAINED),
-    "0", "100", std::nullopt, now }));
-
-    // =========================================================================
-    // LIMITS (limits) - added by #703
-    // The sizes and counts a run or a collection may not exceed. Every entry
-    // here is arrived at from a message that names the setting, which is why
-    // they deserve one shelf instead of hiding among infrastructure. All but
-    // one refuse the oversized input outright and say which knob refused it;
-    // `maxDesignResponseBodyBytes` (issue #1157) is the exception and belongs
-    // here anyway, because what it bounds is a read in flight rather than
-    // anything kept on disk - it stops reading at the bound and the response
-    // viewer says the rest was never read. A bound on what is *stored* is the
-    // shelf above.
-    // =========================================================================
-
-    upsert_config (ConfigEntry{ "maxScenarioSteps",
-    std::to_string (vayu::core::constants::scenario::MAX_STEPS), "integer", "Max Scenario Steps",
-    "Largest number of requests one collection run may resolve to. The whole "
-    "sequence is composed before the first send and held in memory for the "
-    "run, and a load-mode scenario allocates a latency histogram per step, so "
-    "this bounds memory rather than expressing a preference. A collection that "
-    "resolves to more steps is rejected outright, never silently truncated.",
-    "limits", std::to_string (vayu::core::constants::scenario::MAX_STEPS), "1",
-    "10000", std::nullopt, now });
-
-    upsert_config (ConfigEntry{ "maxScenarioDataRows",
-    std::to_string (vayu::core::constants::scenario::MAX_DATA_ROWS), "integer", "Max Scenario Data Rows",
-    "Largest data set one collection run may carry. The app parses the CSV or "
-    "JSON file and sends the rows on the run payload - the engine never reads "
-    "a file from disk - so this bounds how big that payload may get. A larger "
-    "data set is rejected rather than truncated.",
-    "limits", std::to_string (vayu::core::constants::scenario::MAX_DATA_ROWS),
-    "1", "1000000", std::nullopt, now });
-
-    upsert_config (unit ("bytes") (ConfigEntry{ "maxScenarioDataBytes",
-    std::to_string (vayu::core::constants::scenario::MAX_DATA_BYTES), "integer", "Max Scenario Data Size",
-    "Largest data set one collection run may carry, measured over its JSON. "
-    "The row limit alone does not bound the payload - one row is free to hold "
-    "a "
-    "megabyte in a single cell - and the HTTP body cap above this would drop "
-    "the connection instead of explaining itself. A larger data set is "
-    "rejected "
-    "with a message naming this setting.",
-    "limits", std::to_string (vayu::core::constants::scenario::MAX_DATA_BYTES),
-    "1024", "104857600", std::nullopt, now }));
-
-    upsert_config (unit ("bytes") (
-    keywords ({ "swagger" }) (ConfigEntry{ "maxSpecDocumentBytes",
-    std::to_string (vayu::core::constants::spec_document::MAX_BYTES), "integer", "Max OpenAPI Document Size",
-    "Largest OpenAPI document one collection may bind. The document is stored "
-    "verbatim and parsed back by every feature that reads it, so this bounds "
-    "both the row and that parse. A larger document is rejected with a message "
-    "naming this setting, never stored truncated.",
-    "limits", std::to_string (vayu::core::constants::spec_document::MAX_BYTES),
-    "1024", "104857600", std::nullopt, now })));
-
-    upsert_config (unit ("bytes") (ConfigEntry{ "maxResponseBodyBytes",
-    std::to_string (vayu::core::constants::event_loop::MAX_RESPONSE_BODY_BYTES),
-    "integer", "Max Load-Test Response Body",
-    "Largest response body a single load-test request will read into "
-    "memory. A larger response fails that request with an error instead of "
-    "being buffered, so load testing a big download or a streaming endpoint "
-    "cannot exhaust memory - every in-flight request holds its own body. "
-    "Design-mode sends are not affected: they read their own bound, "
-    "Max Design Response Body, and keep what they read instead of failing.",
-    "limits", std::to_string (vayu::core::constants::event_loop::MAX_RESPONSE_BODY_BYTES),
-    "1024",       // 1KB
-    "1073741824", // 1GB
-    std::nullopt, now }));
-
-    upsert_config (unit ("bytes") (ConfigEntry{ "maxDesignResponseBodyBytes",
-    std::to_string (vayu::core::constants::http::MAX_DESIGN_RESPONSE_BODY_BYTES),
-    "integer", "Max Design Response Body",
-    "Largest response body a single Send - or one step of a collection "
-    "run - reads into memory. A larger response stops being read at this "
-    "point: the response viewer shows what arrived and says the rest was "
-    "not read, rather than the whole of it being buffered by the engine and "
-    "held again by the app. Re-sending reads the same amount, so raise this "
-    "to see more of a big download.",
-    "limits", std::to_string (vayu::core::constants::http::MAX_DESIGN_RESPONSE_BODY_BYTES),
-    "1024",       // 1KB
-    "1073741824", // 1GB
-    std::nullopt, now }));
-
-    upsert_config (unit ("bytes") (ConfigEntry{ "maxElementBodyBytes",
-    std::to_string (vayu::core::constants::elements::MAX_BODY_BYTES), "integer", "Max Element Body",
-    "Largest response body an extract.json or assert.jsonpath element will "
-    "parse as JSON. A larger response is not parsed at all: every "
-    "JSON-reading element on that step reports skipped, with the reason, "
-    "instead of paying for - or failing on - a parse of an oversized body. "
-    "The one shared parse is reused by every such element on the same step, "
-    "so this is a per-step cost, not a per-element one.",
-    "limits", std::to_string (vayu::core::constants::elements::MAX_BODY_BYTES),
-    "1024",       // 1KB
-    "1073741824", // 1GB
-    std::nullopt, now }));
-
-    upsert_config (advanced (keywords ({ "infinite loop" }) (
-    ConfigEntry{ "maxStepsPerIteration", "0", "integer", "Max Steps Per Iteration",
-    "How many requests one iteration of a collection run may send before it is "
-    "stopped. It exists because a script can redirect the sequence with "
-    "pm.execution.setNextRequest, and two steps pointing at each other would "
-    "otherwise run forever. 0 derives the limit from the collection - ten "
-    "times its request count, never fewer than 100.",
-    "limits", "0", "0", "1000000", std::nullopt, now })));
-
-    // =========================================================================
-    // SCRIPTING ENVIRONMENT (scripting_sandbox)
-    // The QuickJS sandbox a pre- or post-request script runs in: its deadline,
-    // its memory and stack, and whether console output is collected.
-    // =========================================================================
-
-    upsert_config (
-    unit ("ms") (keywords ({ "runaway" }) (ConfigEntry{ "scriptTimeout",
-    std::to_string (vayu::core::constants::script_engine::TIMEOUT_MS), "integer", "Script Execution Timeout",
-    "How long a pre- or post-request script may run. A script "
-    "that exceeds this is aborted and reported as an error, so an infinite "
-    "loop "
-    "cannot hang the engine. 0 disables the limit, which is not recommended.",
-    "scripting_sandbox", std::to_string (vayu::core::constants::script_engine::TIMEOUT_MS),
-    "0", "60000", std::nullopt, now })));
-
-    upsert_config (
-    keywords ({ "debug", "print" }) (ConfigEntry{ "scriptEnableConsole",
-    vayu::core::constants::script_engine::ENABLE_CONSOLE ? "true" : "false", "boolean", "Enable Script Console",
-    "Makes console.log() available inside scripts, with its output shown in "
-    "the "
-    "response viewer. Turn it off during a load test, where the writes cost "
-    "throughput for output nobody reads.",
-    "scripting_sandbox", vayu::core::constants::script_engine::ENABLE_CONSOLE ? "true" : "false",
-    std::nullopt, std::nullopt, std::nullopt, now }));
-
-    upsert_config (unit ("bytes") (
-    keywords ({ "ram", "oom" }) (ConfigEntry{ "scriptMemoryLimit",
-    std::to_string (vayu::core::constants::script_engine::MEMORY_LIMIT), "integer", "Script Memory Limit",
-    "Largest heap one script execution may allocate before it is aborted. "
-    "Raise "
-    "it only for a script that processes very large data structures.",
-    "scripting_sandbox", std::to_string (vayu::core::constants::script_engine::MEMORY_LIMIT),
-    "1048576",   // 1MB
-    "268435456", // 256MB
-    std::nullopt, now })));
-
-    upsert_config (advanced (unit ("bytes") (keywords ({ "recursion" }) (ConfigEntry{ "scriptStackSize",
-    std::to_string (vayu::core::constants::script_engine::STACK_SIZE), "integer", "Script Stack Size",
-    "Depth of the call stack one script execution may use. Raise it only for a "
-    "script that recurses deeply enough to overflow.",
-    "scripting_sandbox", std::to_string (vayu::core::constants::script_engine::STACK_SIZE),
-    "65536",   // 64KB
-    "1048576", // 1MB
-    std::nullopt, now }))));
+    auto existing               = impl_->storage.get_all<ConfigEntry> ();
+    const size_t existing_count = existing.size ();
+    const auto now              = config_seeds::now_ms ();
+
+    config_seeds::ConfigSeeder seed (
+    std::move (existing), known_config_keys_,
+    [this] (const ConfigEntry& entry) { impl_->storage.replace (entry); },
+    [this] (const std::string& key) {
+        impl_->storage.remove_all<ConfigEntry> (where (c (&ConfigEntry::key) == key));
+    });
+
+    config_seeds::seed_general (seed, now);
+    config_seeds::seed_network (seed, now);
+    config_seeds::seed_services (seed, now);
+    config_seeds::seed_observability (seed, now);
+    config_seeds::seed_data_retention (seed, now);
+    config_seeds::seed_limits (seed, now);
+    config_seeds::seed_scripting (seed, now);
 
     seed_transaction.commit ();
 
-    if (existing.empty ()) {
+    if (existing_count == 0) {
         vayu::utils::log_info ("db", "Seeded default configuration values");
     } else {
         vayu::utils::log_info ("db",
-        "Updated configuration metadata for " +
-        std::to_string (existing.size ()) + " existing entries");
+        "Updated configuration metadata for " + std::to_string (existing_count) + " existing entries");
     }
 }
+
 
 } // namespace vayu::db
